apps/api/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pgvector==0.3.2              # pgvector SQLAlchemy integration

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2

# Google ADK and Gemini
google-adk
google-generativeai

# Protobuf
protobuf==5.28.2

# Background tasks
celery==5.4.0
redis==5.1.1

# RAG / Knowledge base pipeline
sentence-transformers==3.1.1  # all-MiniLM-L6-v2 embeddings
faiss-cpu==1.8.0              # FAISS index build + export for mobile
langchain
langchain-community
pypdf==4.3.1                  # PDF parsing

# File uploads
python-magic==0.4.27          # MIME type validation on upload
aiofiles==24.1.0              # Async file writes

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Real-time
python-socketio==5.11.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
